pyproject.toml
[project]
name = "ai-devs-3-reloaded-tasks"
version = "0.1.0"
description = "API solutions for AI Devs 3 Reloaded course"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
